+++ Sources/ChessKitEngine/EngineType.swift
@@ -5,13 +5,21 @@
 
 import ChessKitEngine_ObjC
 
-extension EngineType {
+public enum EngineType: Int {
+    
+    case stockfish
+    
+    /// Internal mapping from Swift to Obj-C type.
+    var objc: EngineType_ObjC {
+        switch self {
+        case .stockfish:    return .stockfish
+        }
+    }
     
     /// The user-readable name of the engine.
     public var name: String {
         switch self {
         case .stockfish:    return "Stockfish"
-        @unknown default:   return "Unknown"
         }
     }
     
@@ -19,7 +27,6 @@ extension EngineType {
     public var version: String {
         switch self {
         case .stockfish:    return "15.1"
-        @unknown default:   return ""
         }
     }