+++ Sources/ChessKitEngine_ObjC/EngineMessenger/EngineType_ObjC.h
@@ -1,0 +1,13 @@
+//
+//  EngineType_ObjC.h
+//  ChessKitEngine
+//
+
+#ifndef EngineType_h
+#define EngineType_h
+
+typedef NS_ENUM(NSInteger, EngineType_ObjC) {
+    EngineTypeStockfish
+};
+
+#endif /* EngineType_ObjC_h */