+++ Sources/ChessKitEngine_ObjC/include/ChessKitEngine_ObjC.h
@@ -7,7 +7,7 @@
 #define ChessKitEngine_ObjC_h
 
 #include "../EngineMessenger/EngineMessenger.h"
-#include "../EngineMessenger/EngineType.h"
+#include "../EngineMessenger/EngineType_ObjC.h"
 #include "../Logging/Logging.h"
 
 #endif /* ChessKitEngine_ObjC_h */