+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+# ChessKitEngine 0.1.2
+Released Friday, April 14, 2023.
+
+* Hide internal Obj-C and C++ targets
+    * Created separate Swift and Obj-C `EngineType` enums
+    * Package user should only need to import `ChessKitEngine` now
+
 # ChessKitEngine 0.1.1
 Released Friday, April 14, 2023.
 
-* Fix build issue related to missing ChessKitEngine_Cxx target
+* Fix build issue related to missing `ChessKitEngine_Cxx` target
         
 # ChessKitEngine 0.1.0
 Released Friday, April 14, 2023.